Clean up TableViewModel
