+++ Authenticator/Classes/TokenForm.swift
@@ -9,6 +9,15 @@
 import OneTimePasswordLegacy
 
 @objc
+protocol TableViewModel {
+    var title: String { get }
+    var numberOfSections: Int { get }
+    func numberOfRowsInSection(section: Int) -> Int
+    func cellForRowAtIndexPath(indexPath: NSIndexPath) -> UITableViewCell?
+    func viewForHeaderInSection(section:Int) -> UIView?
+}
+
+@objc
 protocol TokenForm: TableViewModel {
     weak var presenter: TokenFormPresenter? { get set }
